// src/data/node-types/model-cars/types/ModelCarNode.ts
import type {DataNodeType} from "../../../types/DataNodeType"

export type ModelCarNode = {
    type: DataNodeType.MODEL_CAR
    data: {
        id: number
        name: string
        product_code?: string
        release_year?: number
        scale?: string
        series?: string
        created_at: string
        updated_at: string
    }
}